+++ lib/core/__init__.py
@@ -4,3 +4,4 @@ from .fuzzer import *  # noqa: F401
 from .path import *  # noqa: F401
 from .report_manager import *  # noqa: F401
 from .raw import *  # noqa: F401
+from .advanced import *