+++ lib/output/cli_output.py
@@ -89,7 +89,7 @@ class CLIOutput(object):
         self.lastInLine = False
         sys.stdout.flush()
 
-    def statusReport(self, path, response, full_url, addedToQueue, templates):
+    def statusReport(self, path, response, full_url, addedToQueue, signatures):
         contentLength = None
         status = response.status
 
@@ -139,9 +139,8 @@ class CLIOutput(object):
         if addedToQueue:
             message += "     (Added to queue)"
 
-        if templates:
-            for template in templates:
-                message += "\n==> Template: {0}".format(template)
+        for signature in signatures:
+            message += "\n==> Signature found: {0}".format(signature)
 
         with self.mutex:
             self.newLine(message)